requirements.txt
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.1.0
